add fiture admin and superadmin

+++ src/app/actions/delete-property.ts
@@ -1,20 +1,18 @@
 "use server";
 
 import { prisma } from "@/lib/prisma";
+import { requireSuperadmin } from "@/lib/require-superadmin";
 import { revalidatePath } from "next/cache";
 
-export async function deleteProperty(
-  id: string
-) {
+export async function deleteProperty(id: string) {
+  await requireSuperadmin();
   await prisma.property.delete({
     where: {
       id,
     },
   });
 
-  revalidatePath(
-    "/dashboard/properties"
-  );
+  revalidatePath("/dashboard/properties");
 
   return {
     success: true,
